move test pkg
