+++ providers/flagd/src/test/java/dev/openfeature/contrib/providers/flagd/strategy/TracedResolvingTest.java
@@ -1,4 +1,4 @@
-package dev.openfeature.contrib.providers.flagd;
+package dev.openfeature.contrib.providers.flagd.strategy;
 
 
 import com.google.protobuf.Message;
